requirements.txt
google.generativeai
os
streamlit
numpy
opencv-python
pillow
python-docx

